move storeTools logic into SyncRepository
